Doom: OMG Edition
Damn.

Still can't believe anything runs on this. As a big fan of both 2016 and
Eternal, this is indeed a glorious day.

There's a bug when loading a save, and the key press drops make the game
hard to enjoy reliably, but hell, it does work!

I had to supply an implementation for some rather obscure calls (such as
the now beloved `__divdi3`), and my initial naive attempts didn't quite
satisfy doom. This should explain the added `arith64.c` file; many
thanks to its author, linked within. Perhaps it'd be better to link with
libgcc, but right now, let's not.